# lib 2/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Productivity Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List all habits
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{id}:
    patch:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /habits/{id}/check:
    post:
      operationId: checkHabit
      tags: [habits]
      summary: Mark habit as done for today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated habit with streak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"

  # ── FINANCES ───────────────────────────────────────────────────────────────
...
  /pdf/parse-tasks:
    post:
      operationId: parsePdfTasks
      tags: [pdf]
      summary: Accept extracted PDF text and return parsed tasks for review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfTextInput"
      responses:
        "200":
          description: Parsed task suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PdfTaskSuggestion"
  /pdf/confirm-tasks:
    post:
      operationId: confirmPdfTasks
      tags: [pdf]
      summary: Confirm and save selected parsed tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfTaskConfirmInput"
      responses:
        "201":
          description: Saved tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, missed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
        dueTime:
          type: ["string", "null"]
        tagId:
          type: ["integer", "null"]
        tagName:
          type: ["string", "null"]
        tagColor:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        fromPdf:
          type: boolean
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
        dueTime:
          type: string
        tagId:
          type: integer
        fromPdf:
          type: boolean
...
    Habit:
      type: object
      required: [id, name, streak, currentStreak, level, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        streak:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        level:
          type: integer
        completedToday:
          type: boolean
        lastChecked:
          type: ["string", "null"]
        tagId:
          type: ["integer", "null"]
        createdAt:
          type: string
    HabitInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        tagId:
          type: integer
    HabitUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        tagId:
          type: integer

    # ── Finance schemas ───────────────────────────────────────────────────────
...
    PdfTextInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
    PdfTaskSuggestion:
      type: object
      required: [title, confidence]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
        confidence:
          type: number
    PdfTaskConfirmInput:
      type: object
      required: [tasks]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskInput"